libtorc-core: Update TorcBonjour to respond to network status.
The bonjour service will be suspended when the network is unavailable
(as already happens when sleeping) and resumed when the network is
available again. It is suspended at startup if the network is
unavailable and all requests are queued. When resuming from sleep, it is
resumed only when the network is back up again (and not just when
woken).

Probably still requires additional work, as it is not clear what action
the mDNS server is actually taking internally.