libtorc-core: Add initial services framework to extend HTTP supprt.
- this is not complete.
- all compound QVariant return types are supported but QVariantList's
with different types will produce an error in the returned results.
- no support yet for custom (QObject derived) types or WSDL.

QObjects inheriting from TorcHTTPService will have their pubic slots
exposed via the HTTP server. As the API is designed to extend existing
core objects, there may be slots that are not suited to public
'exposure' and these can be excluded by adding a coma separated list of
'blacklisted' functions to the constructor.

By default, services are listed under the /services directory and some
simple, human readable help can be displayed using the help method (e.g.
http://torcclient:port/services/power/Help).

Currently available services are listed if a request is sent to the root
/services directory.