libtorc-baseui: Add a secondary UIEffect to UIWidget
- this initially replaces/extends the reflection effect. By using a full
UIEffect for the reflection it becomes far more flexible and can be used
for many more effects. Initially just use it to add alpha to the
reflection - but in the future could be use for drop shadows etc.

- theme xml wise, just add a nested effect element.

NB adding a reflection attribute to both the main and secondary effects
will likely lead to heartache and tears.