libtorc-qml: Fix TorcQMLMediaElement threading behaviour
- the Qt render thread doesn't appear be a full QThread with an event
loop.
- TorcPlayer needs to reside in the same thread as TorcPlayerInterface
to ensure certain slots are guaranteed to be thread safe. Furthermore,
other QRunnable's (e.g via the HTTP interface) will also invoke
TorcPlayer slots and we can't use signals from a QRunnable to ensure
that thread safety.

So move the TorcPlayer instance in TorcQMLMediaElement into the main
thread and move the TorcSGVideoProvider out of TorcSGVideoPlayer and
ensure it is created and destroyed from within the Qt render thread.

N.B. the main thread is blocked during calls to
TorcQMLMediaElement::updatePaintNode, so it is safe to access the player
instance from within that method.

Fixes player HTTP interface.